refactor(mentor): 파일구조 수정
