--- routes/deleteProduct.js
@@ -1,22 +1,0 @@
-import { Router } from "express";
-import Product from "../models/product.js";
-import asyncHandler from "./asyncHandler.js";
-const router = Router();
-
-// 상품 삭제 API
-router.delete(
-  "/:id",
-  asyncHandler(async (req, res) => {
-    try {
-      const product = await Product.findByIdAndDelete(req.params.id);
-      if (!product) {
-        return res.status(404).json({ error: "Product not found" });
-      }
-      res.status(204).send();
-    } catch (error) {
-      res.status(400).json({ error: error.message });
-    }
-  })
-);
-
-export default router;
