+++ routes/products.js
@@ -1,14 +1,11 @@
 import express from "express";
-import createProduct from "./createProduct.js";
-import getProducts from "./getProducts.js";
-import updateProduct from "./updateProduct.js";
-import deleteProduct from "./deleteProduct.js";
+import productService from "../service/productService.js";
 
 const router = express.Router();
-router.post("/", createProduct);
-router.get("/", getProducts);
-router.get("/:id", getProducts);
-router.patch("/:id", updateProduct);
-router.delete("/:id", deleteProduct);
+router.post("/", productService);
+router.get("/", productService);
+router.get("/:id", productService);
+router.patch("/:id", productService);
+router.delete("/:id", productService);
 
 export default router;
